Making room for index.php to do routing
